Move and test PaginatedUsage directive
Move to a dedicated module and add unit tests.
