+++ src/globus_sdk/_sphinxext/__init__.py
@@ -12,26 +12,9 @@ from .custom_directives import (
     ExpandTestingFixture,
     ExternalDocLink,
     ListKnownScopes,
+    PaginatedUsage,
 )
 from .roles import extdoclink_role
-
-
-class PaginatedUsage(AddContentDirective):
-    has_content = False
-    required_arguments = 1
-    optional_arguments = 0
-
-    def gen_rst(self):
-        yield "This method supports paginated access. "
-        yield "To use the paginated variant, give the same arguments as normal, "
-        yield "but prefix the method name with ``paginated``, as in"
-        yield ""
-        yield ".. code-block::"
-        yield ""
-        yield f"    client.paginated.{self.arguments[0]}(...)"
-        yield ""
-        yield "For more information, see"
-        yield ":ref:`how to make paginated calls <making_paginated_calls>`."
 
 
 class CopyParams(AddContentDirective):
